Finalize month punit (#20)
* create Month LayOut

* DataBase Update

* Update Api And Services

* Create Month Function Update

* weekdayVSSegment Table

* migration WeekdaySegmentConfig

* Create Table  DistributionChannelVSsegmentConfig

* Comment Some Code

* PriceDecision Calculation

* Calculation IncomeState

* Comit All

* Create IdealRatingAttributeWeightConfig

* CustomerRowRatting,WeightAttributeRatting

* RoomAllocation Calculation

* Room Allocation Calculation

* SoldRoomByChannel

* BalanceSheet,IncomeState For Calculation

* calculation Function

* Create Ranking Tables

* Calculation MOnth

* calculation Function Update

* sleep function remove

* Create Api for Calculation List

* Some Calculation Function Update

* CreateCalculation Component

* Some Change Calculation ,Create month

* Create month FUNCTION UPDATE

* Price Decision Data Update Function

* Some Function update Calculation

* ClientSide Createmonth update

* Calculation Function Update

* 1.Finalize Button Function update
2. And Other Function Update

* Changes in Ui
1. Create month And Calculation Month

* Some Function update for Check month is Completed

* Month Grid Show Status

* month list show by classId

* month list Api load data ClassId Wise And Show Class Name

* Create Month Use Madaster

* Calculate Month with Mapster

* bug fix Create month and calculation

* update Table Columns int to decimal

* AttributeDecisionDto property update

* Some Table Columns Update

* function Update after Columns Value

* pricedecision,RoomAllocation etc some Table Columns Update

* Some Change AfterTable Columns Update

* Function Change single Query Function not use .
use direct Query

* function update return single decimal Value

* Function Update

---------

+++ Hotel-backend/Database/Domain/CustomerRawRating.cs
@@ -17,7 +17,7 @@ public class CustomerRawRating
     public int GroupID { get; set; }
     public string Attribute { get; set; }
     public string Segment { get; set; }
-    public int RawRating { get; set; }
+    public decimal RawRating { get; set; }
     public virtual Month Month { get; set; }
 
 }
