+++ Hotel-backend/Database/Domain/Goal.cs
@@ -15,26 +15,26 @@ public class Goal
     public int MonthID { get; set; }
     public int QuarterNo { get; set; }
     public int GroupID { get; set; }
-    public int OccupancyM { get; set; }
-    public int OccupancyY { get; set; }
-    public int RoomRevenM { get; set; }
-    public int RoomRevenY { get; set; }
-    public int TotalRevenM { get; set; }
-    public int TotalRevenY { get; set; }
-    public int ShareRoomM { get; set; }
-    public int ShareRoomY { get; set; }
-    public int ShareRevenM { get; set; }
-    public int ShareRevenY { get; set; }
-    public int RevparM { get; set; }
-    public int RevparY { get; set; }
-    public int ADRM { get; set; }
-    public int ADRY { get; set; }
-    public int YieldMgtM { get; set; }
-    public int YieldMgtY { get; set; }
-    public int MgtEfficiencyM { get; set; }
-    public int MgtEfficiencyY { get; set; }
-    public int ProfitMarginM { get; set; }
-    public int ProfitMarginY { get; set; }
+    public decimal OccupancyM { get; set; }
+    public decimal OccupancyY { get; set; }
+    public decimal RoomRevenM { get; set; }
+    public decimal RoomRevenY { get; set; }
+    public decimal TotalRevenM { get; set; }
+    public decimal TotalRevenY { get; set; }
+    public decimal ShareRoomM { get; set; }
+    public decimal ShareRoomY { get; set; }
+    public decimal ShareRevenM { get; set; }
+    public decimal ShareRevenY { get; set; }
+    public decimal RevparM { get; set; }
+    public decimal RevparY { get; set; }
+    public decimal ADRM { get; set; }
+    public decimal ADRY { get; set; }
+    public decimal YieldMgtM { get; set; }
+    public decimal YieldMgtY { get; set; }
+    public decimal MgtEfficiencyM { get; set; }
+    public decimal MgtEfficiencyY { get; set; }
+    public decimal ProfitMarginM { get; set; }
+    public decimal ProfitMarginY { get; set; }
     public virtual Month Month { get; set; }
 
 }
